Refactor abacus creation
