+++ src/graph_broker/util.rs
@@ -19,192 +19,6 @@ use super::{abacus::GraphMask, graph::GraphStorage, ItemId, Orientation, PathSeg
 
 const CHUNK_SIZE: usize = 8192;
 
-pub fn parse_gfa_paths_walks_multiple<R: Read>(
-    data: &mut BufReader<R>,
-    graph_mask: &GraphMask,
-    graph_storage: &GraphStorage,
-    count_types: &Vec<CountType>,
-) -> (
-    Vec<ItemTable>,
-    Vec<Option<ActiveTable>>,
-    Option<IntervalContainer>,
-    HashMap<PathSegment, (u32, u32)>,
-) {
-    log::info!("parsing path + walk sequences");
-    let mut item_tables =
-        vec![ItemTable::new(graph_storage.path_segments.len()); count_types.len()];
-
-    let (mut subset_covered_bps, mut exclude_tables, include_map, exclude_map) =
-        graph_mask.load_optional_subsetting_multiple(graph_storage, count_types);
-
-    let mut num_path = 0;
-    let complete: Vec<(usize, usize)> = vec![(0, usize::MAX)];
-    let mut paths_len: HashMap<PathSegment, (u32, u32)> = HashMap::new();
-
-    let mut buf = vec![];
-    let timer = Instant::now();
-    while data.read_until(b'\n', &mut buf).unwrap_or(0) > 0 {
-        if buf[0] == b'P' || buf[0] == b'W' {
-            let (path_seg, buf_path_seg) = match buf[0] {
-                b'P' => parse_path_identifier(&buf),
-                b'W' => parse_walk_identifier(&buf),
-                _ => unreachable!(),
-            };
-
-            log::debug!("processing path {}", &path_seg);
-
-            let include_coords = if graph_mask.include_coords.is_none() {
-                &complete[..]
-            } else {
-                match include_map.get(&path_seg.id()) {
-                    None => &[],
-                    Some(coords) => {
-                        log::debug!(
-                            "found include coords {:?} for path segment {}",
-                            &coords[..],
-                            &path_seg.id()
-                        );
-                        &coords[..]
-                    }
-                }
-            };
-            let exclude_coords = if graph_mask.exclude_coords.is_none() {
-                &[]
-            } else {
-                match exclude_map.get(&path_seg.id()) {
-                    None => &[],
-                    Some(coords) => {
-                        log::debug!(
-                            "found exclude coords {:?} for path segment {}",
-                            &coords[..],
-                            &path_seg.id()
-                        );
-                        &coords[..]
-                    }
-                }
-            };
-
-            let (start, end) = path_seg.coords().unwrap_or((0, usize::MAX));
-
-            // do not process the path sequence if path is neither part of subset nor exclude
-            if graph_mask.include_coords.is_some()
-                && !intersects(include_coords, &(start, end))
-                && !intersects(exclude_coords, &(start, end))
-            {
-                log::debug!("path {} does not intersect with subset coordinates {:?} nor with exclude coordinates {:?} and therefore is skipped from processing",
-                    &path_seg, &include_coords, &exclude_coords);
-
-                // update prefix sum
-                for item_table in &mut item_tables {
-                    item_table.id_prefsum[num_path + 1] += item_table.id_prefsum[num_path];
-                }
-
-                num_path += 1;
-                buf.clear();
-                continue;
-            }
-
-            let mut indices: HashMap<CountType, Vec<usize>> = HashMap::new();
-            for (i, count_type) in count_types
-                .iter()
-                .map(|c| match c {
-                    CountType::Node => &CountType::Bp,
-                    count => count,
-                })
-                .enumerate()
-            {
-                if let Some(entry) = indices.get_mut(count_type) {
-                    (*entry).push(i);
-                } else {
-                    indices.insert(*count_type, vec![i]);
-                }
-            }
-            indices.into_iter().for_each(|(count, is)| {
-                if count != CountType::Edge
-                    && (graph_mask.include_coords.is_none()
-                        || is_contained(include_coords, &(start, end)))
-                        && (graph_mask.exclude_coords.is_none()
-                            || is_contained(exclude_coords, &(start, end)))
-                {
-                    log::debug!("path {} is fully contained within subset coordinates {:?} and is eligible for full parallel processing", path_seg, include_coords);
-                    let mut none = None;
-                    let ex: Vec<&mut Option<ActiveTable>> = if exclude_coords.is_empty() {
-                        vec![&mut none]
-                    } else {
-                        exclude_tables.iter_mut().enumerate().filter(|(i, _)| is.contains(i)).map(|(_, e)| e).collect()
-                    };
-                    let (num_added_nodes, bp_len) = match buf[0] {
-                        b'P' => parse_path_seq_update_tables_multiple(
-                            buf_path_seg,
-                            graph_storage,
-                            &mut item_tables[is[0]],
-                            ex,
-                            num_path,
-                        ),
-                        b'W' => parse_walk_seq_update_tables_multiple(
-                            buf_path_seg,
-                            graph_storage,
-                            &mut item_tables[is[0]],
-                            ex,
-                            num_path,
-                        ),
-                        _ => unreachable!(),
-                    };
-                    paths_len.insert(path_seg.clone(), (num_added_nodes, bp_len));
-                } else {
-                    let sids = match buf[0] {
-                        b'P' => parse_path_seq_to_item_vec(buf_path_seg, graph_storage),
-                        b'W' => parse_walk_seq_to_item_vec(buf_path_seg, graph_storage),
-                        _ => unreachable!(),
-                    };
-                    let mut exclude_tables_red = exclude_tables.iter_mut().enumerate().filter(|(i, _)| is.contains(i)).map(|(_, e)| e).collect();
-                    match count {
-                        CountType::Node | CountType::Bp => {
-                            let (node_len, bp_len) = update_tables_multiple(
-                                &mut item_tables[is[0]],
-                                &mut subset_covered_bps.as_mut(),
-                                exclude_tables_red,
-                                num_path,
-                                graph_storage,
-                                sids,
-                                include_coords,
-                                exclude_coords,
-                                start,
-                            );
-                            paths_len.insert(path_seg.clone(), (node_len as u32, bp_len as u32));
-                        }
-                        CountType::Edge => update_tables_edgecount(
-                            &mut item_tables[is[0]],
-                            &mut exclude_tables_red[0].as_mut(),
-                            num_path,
-                            graph_storage,
-                            sids,
-                            include_coords,
-                            exclude_coords,
-                            start,
-                        ),
-                        CountType::All => unreachable!("inadmissable count type"),
-                    };
-                }
-            });
-            num_path += 1;
-        }
-        buf.clear();
-    }
-    let duration = timer.elapsed();
-    log::info!(
-        "func done; count: {:?}; time elapsed: {:?}",
-        count_types,
-        duration
-    );
-
-    // In case of node/bps, copy first table into the second
-    if item_tables.len() == 2 {
-        item_tables[1] = item_tables[0].clone();
-    }
-    (item_tables, exclude_tables, subset_covered_bps, paths_len)
-}
-
 pub fn parse_gfa_paths_walks<R: Read>(
     data: &mut BufReader<R>,
     graph_mask: &GraphMask,
